添加sql语句，pojo更新

+++ shop/src/main/java/com/janine/pojo/Area.java
@@ -3,59 +3,60 @@ package com.janine.pojo;
 import java.util.Date;
 
 public class Area {
-	private Long areaId;
-	private String areaName;
-	private String areaDesc;
-	private Integer priority;
-	private Date createTime;
-	private Date lastEditTime;
-
-	public Long getAreaId() {
-		return areaId;
-	}
-
-	public void setAreaId(Long areaId) {
-		this.areaId = areaId;
-	}
-
-	public String getAreaName() {
-		return areaName;
-	}
-
-	public void setAreaName(String areaName) {
-		this.areaName = areaName;
-	}
-
-	public String getAreaDesc() {
-		return areaDesc;
-	}
-
-	public void setAreaDesc(String areaDesc) {
-		this.areaDesc = areaDesc;
-	}
-
-	public Integer getPriority() {
-		return priority;
-	}
-
-	public void setPriority(Integer priority) {
-		this.priority = priority;
-	}
-
-	public Date getCreateTime() {
-		return createTime;
-	}
-
-	public void setCreateTime(Date createTime) {
-		this.createTime = createTime;
-	}
-
-	public Date getLastEditTime() {
-		return lastEditTime;
-	}
-
-	public void setLastEditTime(Date lastEditTime) {
-		this.lastEditTime = lastEditTime;
-	}
-
+    // 不用基础类型，因为会赋空值
+    private Integer areaId;
+    private String areaName;
+    private String areaDesc;
+    // 权重，排序用
+    private Integer priority;
+    private Date createTime;
+    private Date lastEditTime;
+    
+    public Integer getAreaId() {
+        return areaId;
+    }
+    
+    public void setAreaId(Integer areaId) {
+        this.areaId = areaId;
+    }
+    
+    public String getAreaName() {
+        return areaName;
+    }
+    
+    public void setAreaName(String areaName) {
+        this.areaName = areaName;
+    }
+    
+    public String getAreaDesc() {
+        return areaDesc;
+    }
+    
+    public void setAreaDesc(String areaDesc) {
+        this.areaDesc = areaDesc;
+    }
+    
+    public Integer getPriority() {
+        return priority;
+    }
+    
+    public void setPriority(Integer priority) {
+        this.priority = priority;
+    }
+    
+    public Date getCreateTime() {
+        return createTime;
+    }
+    
+    public void setCreateTime(Date createTime) {
+        this.createTime = createTime;
+    }
+    
+    public Date getLastEditTime() {
+        return lastEditTime;
+    }
+    
+    public void setLastEditTime(Date lastEditTime) {
+        this.lastEditTime = lastEditTime;
+    }
 }
